Refactor file structure to breakup command types
