+++ src/commands/chat/index.ts
@@ -1,0 +1,3 @@
+export { HelpCommand } from './help-command.js';
+export { InfoCommand } from './info-command.js';
+export { TestCommand } from './test-command.js';
