+++ src/commands/chat/info-command.ts
@@ -8,20 +8,21 @@ import { createRequire } from 'node:module';
 import os from 'node:os';
 import typescript from 'typescript';
 
-import { LangCode } from '../enums/lang-code.js';
-import { Language } from '../models/enum-helpers/language.js';
-import { EventData } from '../models/internal-models.js';
-import { Lang } from '../services/index.js';
-import { InteractionUtils, ShardUtils } from '../utils/index.js';
-import { Command, CommandDeferType } from './index.js';
+import { LangCode } from '../../enums/lang-code.js';
+import { Language } from '../../models/enum-helpers/language.js';
+import { EventData } from '../../models/internal-models.js';
+import { Lang } from '../../services/lang.js';
+import { InteractionUtils } from '../../utils/interaction-utils.js';
+import { ShardUtils } from '../../utils/shard-utils.js';
+import { Command, CommandDeferType } from '../command.js';
 
 const require = createRequire(import.meta.url);
-let Config = require('../../config/config.json');
-let TsConfig = require('../../tsconfig.json');
+let Config = require('../../../config/config.json');
+let TsConfig = require('../../../tsconfig.json');
 
 export class InfoCommand implements Command {
     public metadata: RESTPostAPIChatInputApplicationCommandsJSONBody = {
-        name: Lang.getCom('commands.info'),
+        name: Lang.getCom('chat-commands.info'),
         description: Lang.getRef('commandDescs.info', Lang.Default),
         dm_permission: true,
         default_member_permissions: undefined,
