+++ src/commands/index.ts
@@ -1,4 +1,1 @@
 export { Command, CommandDeferType } from './command.js';
-export { HelpCommand } from './help-command.js';
-export { InfoCommand } from './info-command.js';
-export { TestCommand } from './test-command.js';
